rename: Updated menu bar component name and functionality
The previous 'MenuBar' component has been renamed to 'NavigationMenuBar'. This change was made to better reflect the purpose of the component. The old 'MenuBar' file has been deleted, and a new file for 'NavigationMenuBar' has been created. The import statement in the header.tsx file has also been updated accordingly.